+++ reiche-backend/.gitignore
@@ -26,4 +26,5 @@ go.work.sum
 
 static/
 tmp/
+build/
