--- reiche-backend/internal/handlers/handleutils/types.go
@@ -1,17 +1,0 @@
-package handlers
-
-// Error types
-type ReqErrCode uint8
-const (
-    HashAlreadyExists ReqErrCode = iota
-)
-
-var ReqErrCodeStr = [...]string{
-    "HashAlreadyExists",
-}
-
-type ReqErr struct {
-    Msg string `json:"msg"`
-    Ctx ReqErrCode `json:"ctx"`
-}
-
